fix(perf): move sanitize-html to build-time
