--- dev.log
@@ -1,18 +1,0 @@
-
-> fractalresonance@2.0.0 dev
-> next dev
-
- ⚠ Warning: Next.js inferred your workspace root, but it may not be correct.
- We detected multiple lockfiles and selected the directory of /home/mumega/package-lock.json as the root directory.
- To silence this warning, set `outputFileTracingRoot` in your Next.js config, or consider removing one of the lockfiles if it's not needed.
-   See https://nextjs.org/docs/app/api-reference/config/next-config-js/output#caveats for more information.
- Detected additional lockfiles: 
-   * /home/mumega/frc/package-lock.json
-
- ⚠ Port 3000 is in use by an unknown process, using available port 3003 instead.
-   ▲ Next.js 15.5.9
-   - Local:        http://localhost:3003
-   - Network:      http://5.161.216.149:3003
-
- ✓ Starting...
- ✓ Ready in 1385ms
