+++ public/catalog.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://fractalresonance.com/catalog.schema.json",
   "version": "1.0.0",
-  "generated": "2026-02-02T16:19:37.737Z",
+  "generated": "2026-02-25T09:54:17.852Z",
   "site": {
     "url": "https://fractalresonance.com",
     "name": "Fractal Resonance Cognition",
